refactor: extract security/signing/ sub-package (Pass 2)
First (and likely only) Pass-2 structural move from the visibility-
reduction sweep: move the 5 signing-key implementation types out of
`security/` top-level into a new internal sub-package `security.signing/`,
which Modulith treats as private-by-default — invisible to other modules.

Files moved (5 main + 3 test) via `git mv` with history preservation:

- `JdbcSigningKeyStore` (the `SigningKeyStore` impl)
- `SigningKeyRotationProperties` (`@ConfigurationProperties`)
- `SigningKeyRotationConfig` (`@Configuration` for ShedLock + `@EnableScheduling`)
- `SigningKeyRotationSchedule` (`@Scheduled` cron wrapper)
- `SigningKeyRotator` (the orchestrator)

Plus their three integration / unit tests, relocated to mirror.

Things that stayed at `security/` top-level:
- `SigningKeyStore` (interface) — the cross-module API; kept at the
  module's "front door" because exposing a single type via a
  `@NamedInterface` sub-package is structural noise for no enforcement
  win (per the Pass-1 design grill on (d-light)).
- `SecurityProperties` — reverted to `public` from package-private (was
  Pass 1's only `@ConfigurationProperties`-record-only). It now
  participates in intra-module cross-package use because
  `signing/JdbcSigningKeyStore` reaches for the KEK; the class still
  belongs at the module's "front door" because its purpose is
  general security-wide properties (`limen.security.*` prefix), not
  signing-specific.
- `DefaultSecurityConfig` — unrelated catch-all filter chain config.
- `security/ratelimit/` sub-package — already internal.

Other top-level modules in scope for Pass 2 (auth's persistent-token
trio, identity's bootstrap-admin trio, observability's three filters)
are below the "3+ types sharing a theme" cluster threshold from the
notes file and are deliberately left at top-level.

Updates `architecture.md` §4.15 and `security/package-info.java` to
mention the new sub-package.
