+++ src/main/java/com/stucray/limen/security/signing/JdbcSigningKeyStore.java
@@ -1,5 +1,7 @@
-package com.stucray.limen.security;
+package com.stucray.limen.security.signing;
 
+import com.stucray.limen.security.SecurityProperties;
+import com.stucray.limen.security.SigningKeyStore;
 import com.nimbusds.jose.jwk.JWK;
 import com.nimbusds.jose.jwk.JWKSet;
 import com.nimbusds.jose.jwk.RSAKey;
