+++ src/test/java/com/stucray/limen/security/signing/SigningKeyRotationFlowIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.stucray.limen.security;
+package com.stucray.limen.security.signing;
 
 import com.nimbusds.jose.jwk.JWK;
 import com.nimbusds.jose.jwk.JWKSet;
